requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.2.10
Flask-Mail==0.9.0
Flask-SQLAlchemy==0.16
Flask-WTF==0.9.5
Jinja2==2.7.2
MarkupSafe==0.23
SQLAlchemy==0.7.9
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.4
backports.ssl-match-hostname==3.4.0.2
blinker==1.3
decorator==3.4.0
itsdangerous==0.24
pytz==2014.3
speaklater==1.3
sqlalchemy-migrate==0.7.2
tornado==3.2.1
wsgiref==0.1.2
